util: fix unstable test in the TestNotifyErrorWaitGroupWrapper (#33063)
close pingcap/tidb#33062
